feat: 完成場地列表頁面基本功能
- 添加 TypeScript 路徑別名配置
- 實現場地列表展示
- 實現分頁功能
- 實現篩選功能（場地名稱、價格範圍、容納人數）
- 修正 API 響應處理邏輯

+++ backend/src/main/java/com/ticketbooking/dto/CreateVenueRequest.java
@@ -4,7 +4,6 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class CreateVenueRequest {
@@ -23,5 +22,5 @@ public class CreateVenueRequest {
 
     @NotNull(message = "每小時價格不能為空")
     @Min(value = 0, message = "價格不能為負數")
-    private BigDecimal pricePerHour;
+    private Double pricePerHour;
 }
