+++ backend/src/main/java/com/ticketbooking/dto/VenueDTO.java
@@ -1,7 +1,6 @@
 package com.ticketbooking.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +10,7 @@ public class VenueDTO {
     private String location;
     private String description;
     private Integer capacity;
-    private BigDecimal pricePerHour;
+    private Double pricePerHour;
     private Long merchantId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
